[ZENDNNL] CodeQL/Coverity Fix For 6.1
The LOWOHA examples size their std::vector buffers with int * int
products that are then converted to size_t, so the multiplication is
evaluated in 32 bits and any overflow happens before the widening.
Casting the leftmost operand moves the whole product into 64-bit
arithmetic.

Widened one expression per flagged site, nine in total:
- examples/lowoha_matmul_example.cpp: group and sequential GEMM buffers
  (5 sites)
- examples/lowoha_group_matmul_example.cpp: fp32 group_matmul buffers
  (3 sites)
- examples/lowoha_conv2d_example.cpp: strided conv2d output buffer

No behavioral change. Every dimension involved is a compile-time literal
in the example itself and the largest product is 401,408 elements, so
none of these can overflow as written. They are corrected because the
examples are reference material that users copy into their own code, and
because this same pattern in library code produced real under-allocation
defects.

Note the cast is applied to the leftmost operand rather than around the
whole product; wrapping the product would widen an already-truncated
result and fix nothing.

Clears CodeQL cpp/integer-multiplication-cast-to-long alerts 70 to 78.

+++ examples/lowoha_conv2d_example.cpp
@@ -317,7 +317,8 @@ int run_lowoha_strided_conv2d_test() {
         std::vector<float> input(batch * in_h * in_w * in_c, 1.0f);
         std::vector<float> filter(filter_h * filter_w * in_c * out_c, 0.5f);
         std::vector<float> bias(out_c, 0.0f);
-        std::vector<float> output(batch * out_h * out_w * out_c, 0.0f);
+        std::vector<float> output(
+                static_cast<size_t>(batch) * out_h * out_w * out_c, 0.0f);
 
         // Configure conv_params
         conv_params params;
